requirements.txt
annotated-types==0.7.0
anyio==4.4.0
archspec @ file:///croot/archspec_1697725767277/work
boltons @ file:///C:/ci_311/boltons_1677729932371/work
Brotli @ file:///C:/ci_311/brotli-split_1676435766766/work
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///C:/ci_311/colorama_1676422310965/work
conda @ file:///C:/b/abs_88drkd35d4/croot/conda_1701719561256/work
conda-content-trust @ file:///C:/b/abs_e3bcpyv7sw/croot/conda-content-trust_1693490654398/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///C:/b/abs_b9wp3lr1gn/croot/conda-package-handling_1691008700066/work
conda_package_streaming @ file:///C:/b/abs_6c28n38aaj/croot/conda-package-streaming_1690988019210/work
contourpy==1.3.2
cryptography @ file:///C:/b/abs_e8cnom_zw_/croot/cryptography_1702071486468/work
cycler==0.12.1
distro @ file:///C:/b/abs_a3uni_yez3/croot/distro_1701455052240/work
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna @ file:///C:/ci_311/idna_1676424932545/work
Jinja2==3.1.4
joblib==1.5.1
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kiwisolver==1.4.8
libmambapy @ file:///C:/b/abs_efpsdwt4ya/croot/mamba-split_1698782663578/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
menuinst @ file:///C:/b/abs_e8p75b4m9q/croot/menuinst_1702390332729/work
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.37.1
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas==2.2.3
pillow==11.2.1
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pluggy @ file:///C:/ci_311/pluggy_1676422178143/work
pycosat @ file:///C:/b/abs_31zywn1be3/croot/pycosat_1696537126223/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.2.3
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.1
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
rich==13.7.1
ruamel.yaml @ file:///C:/ci_311/ruamel.yaml_1676439214109/work
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
truststore @ file:///C:/b/abs_55z7b3r045/croot/truststore_1695245455435/work
typer==0.12.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zstandard==0.19.0
